+++ include/Binary/BinaryDeserializer.hpp
@@ -101,7 +101,7 @@ namespace claid
             template<typename T>
             void callEnum(const char* property, T& member)
             {
-                size_t m;
+                int32_t m;
                 this->callInt(property, m);
                 member = static_cast<T>(m);
             }
